Remove the old outline -- there are still some issues in the new outline
