+++ plugins-client/ext.language/language.xml
@@ -2,40 +2,6 @@
     <a:skin id="language_skin"
       src       = "{require('ext/language/language').skin}"/>
 
-    <a:hbox flex="1" visible="false">
-        <a:model id="mdlOutline">
-            <data>
-            </data>
-        </a:model>
-        <a:tree
-          id              = "treeOutline"
-          class           = "searchresults treeOutline"
-          opencloseaction = "onmouseup"
-          model           = "mdlOutline"
-          openadd         = "false"
-          startcollapsed  = "false"
-          nocollapse      = "true"
-          empty-message   = "No outline available"
-          loading-message = "Loading outline. One moment please..."
-          maxheight       = "350"
-          visible         = "false"
-          scrollbar       = "sbShared 32 7 7">
-            <a:each
-              match         = "[entries|entry]"
-            >   
-                <a:caption><?lm
-                    var re = new RegExp("({txtGoToFile.value.substr(1)})", 'i');
-                    var value = [@name];
-                    
-                    value.replace(re, <strong>$1</strong>);
-                    !![@meta] ? "<code style='color:gray;'> : [@meta]</code>" : "";
-                ?></a:caption>
-                
-                <a:icon value="{ide.staticPrefix}/ext/language/img/[@icon].png" />
-            </a:each>
-        </a:tree>
-      </a:hbox>
-
     <a:bar id="barHierarchy" skinset="language_skin" skin="codecomplete" width="300" visible="false">
         <a:vbox anchors="0 0 0 0">
             <a:toolbar>
